style: 코드 정리 및 컴포넌트 분리
